Refactor and split design system components into individual files
